refactor: Entity rename으로 인한 코드 수정
