+++ src/main/java/projectbuildup/saver/domain/ChallengeRecord/entity/ChallengeRecordEntity.java
@@ -1,0 +1,39 @@
+package projectbuildup.saver.domain.ChallengeRecord.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import projectbuildup.saver.domain.challenge.entity.ChallengeEntity;
+import projectbuildup.saver.domain.user.entity.UserEntity;
+import projectbuildup.saver.global.entity.BaseTimeEntity;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Builder
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity(name = "ChallengeRecord")
+public class ChallengeRecordEntity extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    //절약 금액
+    private Long amount;
+
+    //절약 날짜
+    private LocalDateTime savingDate;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name="user_id")
+    private UserEntity user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "challenge_id")
+    private ChallengeEntity challenge;
+
+}
